LPS-57247 Apply the same mechanism to JournalFeedSMDH
